+++ src/core/api/mod.rs
@@ -1,0 +1,3 @@
+pub mod crates;
+pub mod trakt;
+pub mod tv_maze;