+++ src/core/api/tv_maze/series_information.rs
@@ -1,4 +1,5 @@
 use super::episodes_information::Episode;
+pub use super::Rating;
 use super::*;
 use serde::{Deserialize, Serialize};
 use std::hash::{Hash, Hasher};
@@ -234,6 +235,7 @@ pub struct SeriesMainInformation {
     /// with an embedded list of series' episodes.
     #[serde(rename = "_embedded")]
     pub embedded_episode_list: Option<EmbeddedEpisodeList>,
+    // pub externals: ExternalIds,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -308,6 +310,12 @@ pub struct Network {
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Country {
     pub name: String,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct ExternalIds {
+    pub imdb: Option<String>,
+    pub thetvdb: Option<u32>,
 }
 
 pub async fn get_series_main_info_with_url(url: String) -> Result<String, ApiError> {